refactor(core): the display driver functions' API has been reviewed. The data types have been adapted to fit the actual data passed/used/returned. Since the backlight driver has undergone the similar change, the display driver code has been adapted to respect the changes.
[no changelog]

+++ core/embed/io/display/inc/io/display_utils.h
@@ -25,7 +25,7 @@
  * @param end Target backlight level (0-255)
  * @param delay Total duration of the fade effect in milliseconds
  */
-void display_fade(int start, int end, int delay);
+void display_fade(uint8_t start, uint8_t end, int delay);
 
 /**
  * @brief Starts recording the display output to files
